+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Constants/ModuleSettingOptions.php
@@ -16,4 +16,5 @@ class ModuleSettingOptions
     public const ADD_TYPE_TO_CUSTOMPOST_UNION_TYPE = 'add-type-to-custompost-union-type';
     public const DEFAULT_VALUE = 'default-value';
     public const VALUE_FOR_SINGLE_ENDPOINT = 'value-for-single-endpoint';
+    public const VALUE_FOR_ADMIN_CLIENTS = 'value-for-admin-clients';
 }