+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/ModuleResolvers/ModuleResolverTrait.php
@@ -11,4 +11,5 @@ trait ModuleResolverTrait
 {
     use HasMarkdownDocumentationModuleResolverTrait;
     use PluginMarkdownContentRetrieverTrait;
+    use CommonModuleResolverTrait;
 }