+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Services/MenuPages/SettingsMenuPage.php
@@ -109,6 +109,8 @@ class SettingsMenuPage extends AbstractPluginMenuPage
                                     $this->printTextareaField($module, $itemSetting);
                                 } elseif ($type == Properties::TYPE_BOOL) {
                                     $this->printCheckboxField($module, $itemSetting);
+                                } elseif ($type == Properties::TYPE_NULL) {
+                                    $this->printLabelField($module, $itemSetting);
                                 } else {
                                     $this->printInputField($module, $itemSetting);
                                 }
@@ -158,12 +160,16 @@ class SettingsMenuPage extends AbstractPluginMenuPage
             $module = $item['module'];
             $moduleResolver = $this->moduleRegistry->getModuleResolver($module);
             foreach ($item['settings'] as $itemSetting) {
+                $option = $itemSetting[Properties::INPUT] ?? null;
+                // No option => it is a label
+                if ($option === null) {
+                    continue;
+                }
                 $type = $itemSetting[Properties::TYPE] ?? null;
                 /**
                  * Cast type so PHPStan doesn't throw error
                  */
                 $name = (string)$itemSetting[Properties::NAME];
-                $option = $itemSetting[Properties::INPUT];
                 $canBeEmpty = $itemSetting[Properties::CAN_BE_EMPTY] ?? false;
                 /**
                  * If the input is empty, replace with the default
@@ -373,6 +379,20 @@ class SettingsMenuPage extends AbstractPluginMenuPage
                 <input type="checkbox" name="<?php echo self::SETTINGS_FIELD . '[' . $name . ']'; ?>" id="<?php echo $name; ?>" value="1" <?php checked(1, $value); ?> />
                 <?php echo $itemSetting[Properties::DESCRIPTION] ?? ''; ?>
             </label>
+        <?php
+    }
+
+    /**
+     * Display a label
+     *
+     * @param array<string, mixed> $itemSetting
+     */
+    protected function printLabelField(string $module, array $itemSetting): void
+    {
+        ?>
+            <p>
+                <?php echo $itemSetting[Properties::DESCRIPTION] ?? ''; ?>
+            </p>
         <?php
     }