+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/CHANGELOG.md
@@ -140,11 +140,15 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Allow to use unsafe default settings
 - Module "Schema Self Fields"
 - Schema Configuration for the Single Endpoint
+- Exposed the `__schema` field in the ACLs, allowing us to disable introspection
 - Sort fields and connections together, alphabetically
 - In the ACLs and CCLs, display fields for both the `Root` type, and the `QueryRoot`+`MutationRoot` types ([#989](https://github.com/leoloso/PoP/pull/989))
 - The entities from the WordPress data model are not namespaced anymore ([#990](https://github.com/leoloso/PoP/pull/990))
+- Split options into 2 entries, "Default value for Schema Configuration" and "Apply on Admin clients", for the following settings:
+  - Namespacing
+  - Nested mutations
+  - Admin fields
 - Prettyprint GraphQL queries in the module docs
-- Exposed the `__schema` field in the ACLs, allowing us to disable introspection
 
 ### Fixed